network-app/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
PyYAML==5.4.1
kubernetes==29.0.0
aiohttp==3.9.5